// packages/twenty-server/src/engine/core-modules/application/connection-provider/connection-provider-exception-code.enum.ts
export enum ConnectionProviderExceptionCode {
  PROVIDER_NOT_FOUND = 'PROVIDER_NOT_FOUND',
  CLIENT_CREDENTIALS_NOT_CONFIGURED = 'CLIENT_CREDENTIALS_NOT_CONFIGURED',
  TOKEN_EXCHANGE_FAILED = 'TOKEN_EXCHANGE_FAILED',
  REFRESH_FAILED = 'REFRESH_FAILED',
  INVALID_STATE = 'INVALID_STATE',
  INVALID_REQUEST = 'INVALID_REQUEST',
  FORBIDDEN = 'FORBIDDEN',
  INVALID_CONNECTION_PROVIDER_INPUT = 'INVALID_CONNECTION_PROVIDER_INPUT',
  CONNECTION_PROVIDER_NOT_FOUND = 'CONNECTION_PROVIDER_NOT_FOUND',
  CONNECTION_PROVIDER_NAME_ALREADY_EXISTS = 'CONNECTION_PROVIDER_NAME_ALREADY_EXISTS',
  ON_CONNECT_LOGIC_FUNCTION_NOT_FOUND = 'ON_CONNECT_LOGIC_FUNCTION_NOT_FOUND',
  ON_DISCONNECT_LOGIC_FUNCTION_NOT_FOUND = 'ON_DISCONNECT_LOGIC_FUNCTION_NOT_FOUND',
  ON_DISCONNECT_LOGIC_FUNCTION_FAILED = 'ON_DISCONNECT_LOGIC_FUNCTION_FAILED',
}